Track Proc Connector listeners
The current implementation uses reference counting to track whether
proc events should be sent or not. However, it is possible that a
process subscribes several times without unsubscribing. Or it
unsubscribes without being subscribed at all. Both these actions lead
to a situation when the reference counter does not reflect the actual
state.

This patch adds tracking of actual listeners among the reference
counting.
When a process subscribes for proc connector events, then it is added
to a list of listeners and the reference counter is increased.
When a process subscribes the second time in a row, then nothing
happens. The same goes for unsubscription.
Additionally, attempt to remove a listener is made when proc_exit
event is received. Hence all listeners cleaned up properly.

Signed-off-by: Alex Denisov <1101.debian@gmail.com>